requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.12
